requirements.txt
dependency-injector==4.46.0
SQLAlchemy==2.0.38
flask===3.1.0
watchdog==6.0.0
python-dotenv==1.0.0
psycopg2==2.9.10